delete old user settings in favor of one place with all settings
